AISpeech/requirements.txt
vosk==0.3.45
pyaudio==0.2.14
python-dotenv==1.0.0
openai==1.6.1
azure-cognitiveservices-speech==1.33.0
numpy==1.26.3
